Return of divmodular changed to size_t.

+++ math/include/arithmetic.h
@@ -1,12 +1,14 @@
 /**
  * @author Ytalo Ramon
  * @date   12/06/2021
-*/
+ */
+
+#include "stdlib.h"
 
 /**
  * Calculate modular arithmetic.
  * @param a value
  * @param b div_value
  * @return remainder
-*/
-int divmodular(int a, int b);
+ */
+size_t divmodular(int a, int b);
